Removed split landscape layout
