+++ apps/site-landing/src/containers/HomePage/landscape/Background.tsx
