--- apps/site-landing/src/containers/HomePage/landscape/LandscapeGreaterThanMobile/index.tsx
@@ -1,14 +1,0 @@
-import * as React from 'react';
-import { Background } from './Background';
-import { Content } from './Content';
-
-
-export const LandscapeGreaterThanMobile = () => {
-  return (
-    <>
-      <Background />
-      <Content />
-    </>
-  );
-}
-
