--- apps/site-landing/src/containers/HomePage/landscape/styles.module.less
@@ -1,19 +1,0 @@
-#headingWrapper{
-  margin-left: 7%;
-  z-index: 9;
-  .overTheLandscape{
-    position: inherit;
-    z-index: 9;
-  }
-}
-.landscapeContent{
-  min-height: 680px;
-}
-
-
-.landscapeMobile{
-  width: 100%;
-  position: relative;
-  z-index:-1;
-  margin-top: -270px;
-}
