--- apps/site-landing/src/containers/HomePage/landscape/translations.ts
@@ -1,13 +1,0 @@
-import { defineMessages } from 'react-intl';
-
-export const translations = defineMessages({
-  title : {
-      defaultMessage: 'The future is brighter',
-      description: 'site.homepage.landscape.title: The title for the homepage'},
-  description : {
-      defaultMessage: 'Save, invest and spend money with TheCoin, get 100% of benefits and save our Planet.',
-      description: 'site.homepage.landscape.description: The description for the homepage'},
-  button : {
-      defaultMessage: 'Start Now',
-      description: 'site.homepage.landscape.button: Button label for the landscape'}
-  });
